Replace Dao components with Service components in actions add a file named 'serviceContext.xml' for Spring

+++ src/com/createarttechnology/dao/IArticleCommentDao.java
@@ -6,9 +6,9 @@ import com.createarttechnology.domain.ArticleComment;
 
 public interface IArticleCommentDao extends IBaseDao<ArticleComment> {
 	//根据文章ID找到最高楼层（可能中间楼被删除）
-	Integer getMaxFloorByArticleId(Integer articleId);
+	Long getMaxFloorByArticleId(Integer articleId);
 	//根据文章ID统计评论数
-	Integer countCommentsByArticleId(Integer articleId);
+	Long countCommentsByArticleId(Integer articleId);
 	//根据文章ID获得所有评论
 	List<ArticleComment> getAllCommentsByArticleId(Integer articleId);
 	//根据用户ID获得所有评论
